src/test/java/pl/damianszczepanik/jenkins/buildhistorymanager/descriptors/BuildNumberRangeConditionDescriptorTest.java
package pl.damianszczepanik.jenkins.buildhistorymanager.descriptors;

import static org.assertj.core.api.Assertions.assertThat;

import org.junit.Test;

/**
 * @author Damian Szczepanik (damianszczepanik@github)
 */
public class BuildNumberRangeConditionDescriptorTest {

    @Test
    public void getDisplayName_ReturnsDescriptorName() {

        // given
        BuildNumberRangeConditionDescriptor descriptor = new BuildNumberRangeConditionDescriptor();

        // when
        String displayName = descriptor.getDisplayName();

        // then
        assertThat(displayName).isEqualTo("Build number range");
    }
}
